+++ Liersch.Json/JsonReflectionBased/SLJsonConverter.cs
@@ -1,0 +1,20 @@
+﻿//----------------------------------------------------------------------------
+//
+// Copyright © 2013-2019 Dipl.-Ing. (BA) Steffen Liersch
+// All rights reserved.
+//
+// Steffen Liersch
+// Robert-Schumann-Straße 1
+// 08289 Schneeberg
+// Germany
+//
+// Phone: +49-3772-38 28 08
+// E-Mail: S.Liersch@gmx.de
+//
+//----------------------------------------------------------------------------
+
+namespace Liersch.Json
+{
+  // Type Func<T, TResult> isn't used instead due to .NET 3.5 is required for it.
+  public delegate TResult SLJsonConverter<T, TResult>(T arg);
+}